requirements.txt
google-cloud-documentai==2.20.0
google-cloud-firestore==2.13.1
google-cloud-workflows==1.12.1
google-cloud-storage==2.10.0
google-auth==2.23.0
requests==2.31.0
functions-framework==3.4.0
